# feat: Major performance improvements and bug fixes